requirements.txt
anthropic 
pydantic 
instructor 
rich 
streamlit 
fpdf2 
python-dotenv